+++ src/euroeval/dataset_configs/danish.py
@@ -10,6 +10,7 @@ from ..tasks import (
     LA,
     MCRC,
     NER,
+    NLI,
     RC,
     SENT,
     SUMM,
@@ -103,7 +104,6 @@ VALEU_DA_CONFIG = DatasetConfig(
     bootstrap_samples=False,
 )
 
-
 # Unofficial datasets ###
 
 DALA_CONFIG = DatasetConfig(
@@ -176,6 +176,16 @@ WINOGRANDE_DA_CONFIG = DatasetConfig(
     task=COMMON_SENSE,
     languages=[DANISH],
     labels=["a", "b"],
+    unofficial=True,
+)
+
+DANISH_ENTAILMENT_CONFIG = DatasetConfig(
+    name="danish-entailment",
+    pretty_name="The Danish Entailment Dataset",
+    source="EuroEval/danish-entailment",
+    task=NLI,
+    languages=[DANISH],
+    val_split=None,
     unofficial=True,
 )