+++ src/euroeval/tasks.py
@@ -10,6 +10,7 @@ from .prompt_templates import (
     LA_TEMPLATES,
     MULTIPLE_CHOICE_TEMPLATES,
     NER_TEMPLATES,
+    NLI_TEMPLATES,
     RC_TEMPLATES,
     SENT_TEMPLATES,
     SIMPL_TEMPLATES,
@@ -28,6 +29,18 @@ LA = Task(
     default_num_few_shot_examples=12,
     default_max_generated_tokens=NUM_GENERATION_TOKENS_FOR_CLASSIFICATION,
     default_labels=["correct", "incorrect"],
+    uses_logprobs=True,
+)
+
+
+NLI = Task(
+    name="natural-language-inference",
+    task_group=TaskGroup.SEQUENCE_CLASSIFICATION,
+    template_dict=NLI_TEMPLATES,
+    metrics=[m.mcc_metric, m.macro_f1_metric],
+    default_num_few_shot_examples=12,
+    default_max_generated_tokens=NUM_GENERATION_TOKENS_FOR_CLASSIFICATION,
+    default_labels=["entailment", "neutral", "contradiction"],
     uses_logprobs=True,
 )